Move legacy UserRegistrationService
