+++ src/main/java/site/hirecruit/hr/domain/auth/controller/AuthController.kt
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import site.hirecruit.hr.domain.user.service.UserRegistrationService
+import site.hirecruit.hr.domain.user.service.legacy.UserRegistrationService
 import site.hirecruit.hr.domain.auth.dto.AuthUserInfo
 import site.hirecruit.hr.domain.user.dto.UserRegistrationDto
 import site.hirecruit.hr.global.annotation.CurrentAuthUserInfo
