+++ src/main/java/site/hirecruit/hr/domain/user/aop/UserRegistrationAspect.kt
@@ -28,7 +28,7 @@ class UserRegistrationAspect(
     private val securityContextFacade: SecurityContextAccessService
 ) {
 
-    @Pointcut("execution(* site.hirecruit.hr.domain.user.service.UserRegistrationService+.registration(..))")
+    @Pointcut("execution(* site.hirecruit.hr.domain.user.service.legacy.UserRegistrationService+.registration(..))")
     private fun userRegistrationService_registrationMethodPointCut(){}
 
     @Pointcut("execution(* site.hirecruit.hr.domain.user.service.NewUserRegistrationService+.registration(..))")
