+++ src/test/java/site/hirecruit/hr/domain/auth/aop/UserRegistrationAspectTest.kt
@@ -13,7 +13,7 @@ import site.hirecruit.hr.domain.user.dto.UserRegistrationDto
 import site.hirecruit.hr.domain.user.entity.Role
 import site.hirecruit.hr.domain.auth.repository.TempUserRepository
 import site.hirecruit.hr.domain.auth.service.SecurityContextAccessService
-import site.hirecruit.hr.domain.user.service.UserRegistrationService
+import site.hirecruit.hr.domain.user.service.legacy.UserRegistrationService
 import site.hirecruit.hr.domain.test_util.LocalTest
 import site.hirecruit.hr.domain.user.aop.UserRegistrationAspect
 import site.hirecruit.hr.domain.worker.dto.WorkerDto
